requirements.txt
python>=3.10
biopython
pyyaml
tqdm
viennarn